added map info, corrected dist & src structure, fixed video js, fixed mobile accord js
