+++ src/styles/components/advantages.scss
@@ -1,6 +1,5 @@
 .advantages {
     display: flex;
-    // max-width: 72.5%;
     align-self: center;
 
     @include phones {
@@ -19,7 +18,7 @@
     }
 
     &:before {
-        content: url(../img/decor/dots.png);
+        content: url(./img/decor/dots.png);
         position: absolute;
         top: rem(55px);
         right: 0;
