+++ src/styles/components/hamburger.scss
@@ -30,7 +30,7 @@
     width: rem(20px);
     height: rem(20px);
     opacity: 0.7;
-    background: url(../img/icons/close.svg) center center no-repeat;
+    background: url(./img/icons/close.svg) center center no-repeat;
 }
 
 .isActive {
