+++ src/styles/components/radio.scss
@@ -18,7 +18,7 @@
 
             &:before {
                 content: '';
-                background: url(../img/icons/checkmark.svg) center center no-repeat;
+                background: url(./img/icons/checkmark.svg) center center no-repeat;
             }
         }
     }
