+++ src/styles/components/review.scss
@@ -1,5 +1,5 @@
 .review {
-    background: url(../img/decor/blackberry-bush.png) top right no-repeat;
+    background: url(./img/decor/blackberry-bush.png) top right no-repeat;
     background-color: #f8faf9;
     padding: rem(50px) rem(45px) rem(34px);
     overflow: hidden;
