+++ src/styles/components/section-team.scss
@@ -1,6 +1,6 @@
 .section-team {
     position: relative;
-    background: url(../img/decor/branch-right.png) right bottom 200% no-repeat;
+    background: url(./img/decor/branch-right.png) right bottom 200% no-repeat;
 
     @include tablets {
         background-position: right bottom 160%;
@@ -11,7 +11,7 @@
     }
 
     &:after {
-        content: url('../img/decor/team-blueberries.png');
+        content: url('./img/decor/team-blueberries.png');
         position: absolute;
         right: 0;
         bottom: 89%;
