+++ src/styles/fonts.scss
@@ -1,0 +1,111 @@
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-black-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-black-webfont.woff') format('woff');
+    font-weight: 900;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-blackitalic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-blackitalic-webfont.woff') format('woff');
+    font-weight: 900;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-bold-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-bold-webfont.woff') format('woff');
+    font-weight: 700;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-bolditalic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-bolditalic-webfont.woff') format('woff');
+    font-weight: 700;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-extrabold-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-extrabold-webfont.woff') format('woff');
+    font-weight: 800;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-extrabolditalic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-extrabolditalic-webfont.woff') format('woff');
+    font-weight: 800;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-italic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-italic-webfont.woff') format('woff');
+    font-weight: normal;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-light-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-light-webfont.woff') format('woff');
+    font-weight: 300;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-lightitalic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-lightitalic-webfont.woff') format('woff');
+    font-weight: 300;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-medium-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-medium-webfont.woff') format('woff');
+    font-weight: 500;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-mediumitalic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-mediumitalic-webfont.woff') format('woff');
+    font-weight: 500;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-regular-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-regular-webfont.woff') format('woff');
+    font-weight: 400;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-thin-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-thin-webfont.woff') format('woff');
+    font-weight: 100;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'alegreyasans';
+    src: url('./fonts/alegreyasans-thinitalic-webfont.woff2') format('woff2'),
+         url('./fonts/alegreyasans-thinitalic-webfont.woff') format('woff');
+    font-weight: 100;
+    font-style: italic;
+}
